+++ test/acceptance/components/news-article.js
@@ -53,7 +53,15 @@ module.exports = {
   actions() {
     return {
       CLICK_TO_VIEW_STORY: {
-        preconditions(dataStore) {
+        parameters: [
+          {
+            name: 'fakeParameter',
+            generate() {
+                return 'myFakeParameter';
+            },
+          },
+        ],
+        preconditions(fakeParam, dataStore) {
           dataStore.store(`${this.name}HeadingText`, this.getFromPage(`${this.name}.newsArticleHeading.text`));
           dataStore.store(`${this.name}Text`, this.getFromPage(`${this.name}.newsArticleText.text`));
 
@@ -63,12 +71,12 @@ module.exports = {
             ['property', `dataStore`, `${this.name}Text`],
           ];
         },
-        perform(callback) {
+        perform(fakeParam, callback) {
           driver.findElement(By.id(this.options.newsArticleId))
           .click()
           .then(callback, callback);
         },
-        effects(expectedState, dataStore) {
+        effects(fakeParam, expectedState, dataStore) {
           expectedState.stash();
           expectedState.createAndAddComponent({
             type: 'ViewStoryModal',